# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mall DRC API",
        "version": "1.0"
    },
    "paths": {
...
        "/client/{client}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "supprimes les client ayant Id specifier à l'argument",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "client est l'Id du client récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "client n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:client",
                            "read:client"
                        ]
                    }
                ]
            }
        },
...
        "/produit/{produit}": {
            "get": {
                "tags": [
                    "produit"
                ],
                "summary": "affiche un produit spécifique ayant Id specifier à l'argument",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "description": "produit est l'Id du produit récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "visualisation du produit recherché"
                    },
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le produit n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:produit",
                            "read:produit"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "produit"
                ],
                "summary": "supprimes les produit ayant Id specifier à l'argument",
                "operationId": "destroyproduit",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "description": "produit est l'Id du produit récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le produit n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:produit",
                            "read:produit"
                        ]
                    }
                ]
            }
        },
        "/produit/search/nomProduit": {
            "get": {
                "tags": [
                    "produit"
                ],
                "summary": " Affiche Tous produits contenant le nom rechercher dans sa description ou libellé",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "description": "produit designe tous les produits contenant le nom rechercher dans sa description ou libellé",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nom": {
                                                        "type": "string b",
                                                        "example": "ballon"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/detailProduit/{produit}": {
            "get": {
                "tags": [
                    "produit"
                ],
                "summary": "affiches les produits recherché ainsi que le detail de sa categorie",
                "operationId": "detailProduit",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "description": "produit est l'Id du produit récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le produit n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:produit",
                            "read:produit"
                        ]
                    }
                ]
            }
        },
...
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Affiche Tous user se trouvant dans la base de données",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "le nom de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "kofi"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "le nom de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "peterkofi74@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "le mot de passe de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "peter1234"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "veillez vous authentifiez"
                    }
                }
            }
        },
...
    }
}